+++ test/test-deviant-pattern.lua
@@ -3,7 +3,7 @@ local t = require 'tapered'
 package.path = '../src/?.lua;' .. package.path
 local split = require 'split'.split
 
-local s, got
+local s
 
 s = ',foo,bar'
 t.boom(split, {s, '%d*'}, 'Should blow up with deviant pattern.')