+++ CHANGES.md
@@ -1,8 +1,19 @@
 # split version history
 
-## *1.0-0* (July 16, 2015)
+## *1.0.0-1* (July 16, 2015)
 
 + Initial public release
+
+## *1.1.0-1* (July 19, 2015)
+
++ Various small improvements, thanks to [luacheck][luacheck].
++ Test coverage stats, thanks to [luacov][luacov],
+  [luacov-coveralls][luacov-coveralls], and [coveralls][coveralls].
+
+[luacheck]: https://github.com/mpeterv/luacheck
+[luacov]: http://keplerproject.github.io/luacov
+[luacov-coveralls]: https://github.com/moteus/luacov-coveralls
+[coveralls]: https://coveralls.io
 
 Would you reather view the [documentation][d]?